requirements.txt
# Telegram Video Bot Dependencies

# Telegram Bot Framework
python-telegram-bot>=20.0

# Async HTTP client for API calls
aiohttp>=3.9.0

# Configuration management
python-dotenv>=1.0.0

# (Optional) For production deployment
# gunicorn>=21.0.0  # For running with multiple workers
